+++ extensions/clean-keyboard/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [1 hour] - 2024-04-25
 - Added 1 hour duration
 - Added 24 hour duration
+- Added Infinity duration
 
 ## [Initial Version] - 2024-04-25